feat: min bounty of 100

+++ ffa-server/src/main/kotlin/gg/norisk/ffa/server/mechanics/Bounty.kt
@@ -27,7 +27,7 @@ object Bounty {
                     val dbPlayer = PlayerProvider.get(player.uuid)
                     source.sendMessage(Text.translatable("ffa.mechanic.bounty.info", player.name, dbPlayer.bounty))
                 }
-                argument<Int>("bounty", IntegerArgumentType.integer(0)) { bountyToGive ->
+                argument<Int>("bounty", IntegerArgumentType.integer(100)) { bountyToGive ->
                     runsAsync {
                         val player = EntityArgumentType.getPlayer(this, "player")
                         val source = this.source.playerOrThrow
